moved llm logic to llm_assistant
