+++ llm_assistant/models/llm_thread.py
@@ -1,10 +1,10 @@
+import json
 import logging
 
-from odoo import api, fields, models
+from odoo import _, api, fields, models
+from odoo.exceptions import UserError
 
 _logger = logging.getLogger(__name__)
-
-
 
 
 class LLMThread(models.Model):
@@ -211,26 +211,19 @@ class LLMThread(models.Model):
     def get_prepend_messages(self):
         """Hook: return a list of formatted messages to prepend to the conversation."""
         self.ensure_one()
-        messages = super().get_prepend_messages()
 
         if self.prompt_id:
             try:
 
                 # Get messages from the prompt with enhanced context
-                prompt_messages = self.prompt_id.get_messages(self.get_context())
-
-                if prompt_messages:
-                    messages = self.merge_message_lists(prompt_messages, messages)
-                    _logger.info("Added %d messages from prompt '%s'", len(prompt_messages), self.prompt_id.name)
+                return self.prompt_id.get_messages(self.get_context())
 
             except Exception as e:
                 _logger.error("Error getting messages from prompt '%s': %s", self.prompt_id.name, str(e))
                 # Continue without prompt messages rather than failing completely
-                self.message_post(
-                    body=f"Warning: Could not load prompt messages from '{self.prompt_id.name}': {str(e)}"
-                )
+                self.message_post(body=f"Warning: Could not load prompt messages from '{self.prompt_id.name}': {str(e)}")
 
-        return messages
+        return []
 
     def _process_tool_calls(self, assistant_msg):
         """Override _process_tool_calls to implement circuit breaker using assistant's tool_calls_max"""
@@ -276,5 +269,165 @@ class LLMThread(models.Model):
             f"Thread {self.id}: Processing tool calls ({consecutive_tool_calls}/{tool_calls_max} consecutive assistant tool call messages)"
         )
         
-        # Call the parent method for normal processing
-        return (yield from super()._process_tool_calls(assistant_msg))
+        # Call the base tool processing logic directly
+        self.ensure_one()
+        defs = json.loads(assistant_msg.tool_calls or "[]")
+        last_tool_msg = None
+        for tool_def in defs:
+            last_tool_msg = yield from self.message_post_tool_result(tool_def)
+        return last_tool_msg
+
+    # ============================================================================
+    # AI GENERATION LOGIC - Moved from llm_thread module
+    # ============================================================================
+
+    def generate(self, user_message_body, **kwargs):
+        """Main generation method with actual AI intelligence."""
+        self.ensure_one()
+        if self.is_locked:
+            raise UserError(_("This thread is already generating a response."))
+
+        self._lock()
+        try:
+            # Post user message if provided
+            if user_message_body:
+                user_msg = self.message_post(
+                    body=user_message_body,
+                    llm_role='user',
+                    author_id=self.env.user.partner_id.id,
+                    **kwargs
+                )
+                self.env.cr.commit()
+                yield {"type": "message_create", "message": user_msg.message_format()[0]}
+
+            # Get last message to continue from
+            last_message = self._get_last_message_from_history()
+
+            # Continue generation loop
+            while self._should_continue(last_message):
+                if last_message.llm_role in ('user', 'tool'):
+                    # Generate assistant response
+                    last_message = yield from self._generate_assistant_response()
+                elif (last_message.llm_role == 'assistant' and
+                      last_message.tool_calls):
+                    # Process tool calls
+                    last_message = yield from self._process_tool_calls(last_message)
+                else:
+                    break
+
+            return last_message
+        finally:
+            self._unlock()
+
+    def _generate_assistant_response(self):
+        """Generate assistant response using streaming with actual AI intelligence."""
+        message_history = self.get_message_history_recordset()
+        chat_kwargs = {
+            "messages": message_history,
+            "tools": self.tool_ids,
+            "stream": True,
+            "prepend_messages": self.get_prepend_messages(),
+        }
+
+        stream_response = self.sudo().model_id.chat(**chat_kwargs)
+        return (yield from self.message_post_from_stream(
+            stream_response,
+            'assistant',
+            placeholder_text="Thinking..."
+        ))
+
+    def get_message_history_recordset(self, order="ASC", limit=None):
+        """Get LLM messages from the thread using efficient stored field filtering.
+
+        Args:
+            order: Optional order for messages ('ASC' or 'DESC')
+            limit: Optional limit on number of messages to retrieve
+
+        Returns:
+            mail.message recordset containing the LLM messages
+        """
+        self.ensure_one()
+
+        # Use the stored llm_role field for efficient filtering
+        domain = [
+            ("model", "=", self._name),
+            ("res_id", "=", self.id),
+            ("llm_role", "!=", False),  # Only LLM messages
+        ]
+
+        order_clause = "create_date DESC, write_date DESC, id DESC"
+        if order == "ASC":
+            order_clause = "create_date ASC, write_date ASC, id ASC"
+
+        return self.env["mail.message"].search(domain, order=order_clause, limit=limit)
+
+    def _get_last_message_from_history(self):
+        """Get the last LLM message from the message history."""
+        self.ensure_one()
+        result = self.get_message_history_recordset(order="DESC", limit=1)
+        if not result:
+            raise UserError("No LLM message found to process.")
+        return result[0]
+
+    def _should_continue(self, last_message):
+        """Whether to keep looping based on the last message role and content."""
+        if not last_message:
+            return False
+
+        # Use the stored llm_role field for efficient checking
+        if last_message.llm_role in ('user', 'tool'):
+            return True
+
+        if last_message.llm_role == 'assistant' and last_message.tool_calls:
+            return True
+
+        return False
+
+    def _execute_tool(self, tool_name, arguments_str):
+        """Execute a tool and return the result."""
+        self.ensure_one()
+        tool = self.tool_ids.filtered(lambda t: t.name == tool_name)[:1]
+        if not tool:
+            raise UserError(f"Tool '{tool_name}' not found in this thread")
+        arguments = json.loads(arguments_str)
+        return tool.execute(arguments)
+
+    def message_post_tool_result(self, tool_call_def, **kwargs):
+        """Post a tool result message using standard message_post flow."""
+        call_id = tool_call_def.get("id")
+        fn = tool_call_def.get("function", {})
+        name = fn.get("name", "unknown_tool")
+        args = fn.get("arguments")
+
+        # Create placeholder message
+        message = self.message_post(
+            body=f"Executing: {name}…",
+            llm_role='tool',
+            author_id=False,
+            tool_call_id=call_id,
+            tool_call_definition=json.dumps(tool_call_def),
+            tool_name=name,
+            **kwargs
+        )
+        yield {"type": "message_create", "message": message.message_format()[0]}
+
+        # Execute tool and update message
+        try:
+            with self.env.cr.savepoint():
+                result = self.with_context(message=message)._execute_tool(name, args)
+                if not result:
+                    raise UserError(f"No result returned from tool '{name}'")
+
+                # Update message with result
+                message.write({
+                    "tool_call_result": json.dumps(result),
+                    "body": f"Result for {name}"
+                })
+        except Exception as e:
+            message.write({
+                "tool_call_result": json.dumps({"error": str(e)}),
+                "body": f"Error executing {name}"
+            })
+
+        yield {"type": "message_update", "message": message.message_format()[0]}
+        return message
