+++ llm_thread/models/llm_thread.py
@@ -304,185 +304,34 @@ class LLMThread(models.Model):
 
         return message
 
-    def message_post_tool_result(self, tool_call_def, **kwargs):
-        """Post a tool result message using standard message_post flow."""
-        call_id = tool_call_def.get("id")
-        fn = tool_call_def.get("function", {})
-        name = fn.get("name", "unknown_tool")
-        args = fn.get("arguments")
-
-        # Create placeholder message
-        message = self.message_post(
-            body=f"Executing: {name}…",
-            llm_role='tool',
-            author_id=False,
-            tool_call_id=call_id,
-            tool_call_definition=json.dumps(tool_call_def),
-            tool_name=name,
-            **kwargs
-        )
-        yield {"type": "message_create", "message": message.message_format()[0]}
-
-        # Execute tool and update message
-        try:
-            with self.env.cr.savepoint():
-                result = self.with_context(message=message)._execute_tool(name, args)
-                if not result:
-                    raise UserError(f"No result returned from tool '{name}'")
-
-                # Update message with result
-                message.write({
-                    "tool_call_result": json.dumps(result),
-                    "body": f"Result for {name}"
-                })
-        except Exception as e:
-            message.write({
-                "tool_call_result": json.dumps({"error": str(e)}),
-                "body": f"Error executing {name}"
-            })
-
-        yield {"type": "message_update", "message": message.message_format()[0]}
-        return message
-
     # ============================================================================
     # GENERATION FLOW - Refactored to use message_post with roles
     # ============================================================================
 
     def generate(self, user_message_body, **kwargs):
-        """Main generation method using standard message_post flow."""
-        self.ensure_one()
-        if self.is_locked:
-            raise UserError(_("This thread is already generating a response."))
+        """Main generation method - requires llm_assistant module for actual AI generation."""
+        raise UserError(_("Please install the llm_assistant module for actual AI generation."))
 
-        self._lock()
-        try:
-            # Post user message if provided
-            if user_message_body:
-                user_msg = self.message_post(
-                    body=user_message_body,
-                    llm_role='user',
-                    author_id=self.env.user.partner_id.id,
-                    **kwargs
-                )
-                self.env.cr.commit()
-                yield {"type": "message_create", "message": user_msg.message_format()[0]}
-
-            # Get last message to continue from
-            last_message = self._get_last_message_from_history()
-
-            # Continue generation loop
-            while self._should_continue(last_message):
-                if last_message.llm_role in ('user', 'tool'):
-                    # Generate assistant response
-                    last_message = yield from self._generate_assistant_response()
-                elif (last_message.llm_role == 'assistant' and
-                      last_message.tool_calls):
-                    # Process tool calls
-                    last_message = yield from self._process_tool_calls(last_message)
-                else:
-                    break
-
-            return last_message
-        finally:
-            self._unlock()
-
-    def _generate_assistant_response(self):
-        """Generate assistant response using streaming."""
-        message_history = self.get_message_history_recordset()
-        chat_kwargs = {
-            "messages": message_history,
-            "tools": self.tool_ids,
-            "stream": True,
-            "prepend_messages": self.get_prepend_messages(),
+    def get_context(self, base_context=None):
+        context = {
+            **(base_context or {}),
+            'thread_id': self.id,
         }
 
-        stream_response = self.sudo().model_id.chat(**chat_kwargs)
-        return (yield from self.message_post_from_stream(
-            stream_response,
-            'assistant',
-            placeholder_text="Thinking..."
-        ))
+        try:
+            related_record = self.env[self.model].browse(self.res_id)
+            if related_record:
+                context['related_record'] = RelatedRecordProxy(related_record)
+                context['related_model'] = self.model
+                context['related_res_id'] = self.res_id
+            else:
+                context['related_record'] = None
+                context['related_model'] = None
+                context['related_res_id'] = None
+        except Exception as e:
+            _logger.warning("Error accessing related record %s,%s: %s", self.model, self.res_id, e)
 
-    def _process_tool_calls(self, assistant_msg):
-        """Process tool calls from assistant message."""
-        self.ensure_one()
-        defs = json.loads(assistant_msg.tool_calls or "[]")
-        last_tool_msg = None
-        for tool_def in defs:
-            last_tool_msg = yield from self.message_post_tool_result(tool_def)
-        return last_tool_msg
-
-    # ============================================================================
-    # HELPER METHODS - Optimized with stored llm_role field
-    # ============================================================================
-
-    def get_message_history_recordset(self, order="ASC", limit=None):
-        """Get LLM messages from the thread using efficient stored field filtering.
-
-        Args:
-            order: Optional order for messages ('ASC' or 'DESC')
-            limit: Optional limit on number of messages to retrieve
-
-        Returns:
-            mail.message recordset containing the LLM messages
-        """
-        self.ensure_one()
-
-        # Use the stored llm_role field for efficient filtering
-        domain = [
-            ("model", "=", self._name),
-            ("res_id", "=", self.id),
-            ("llm_role", "!=", False),  # Only LLM messages
-        ]
-
-        order_clause = "create_date DESC, write_date DESC, id DESC"
-        if order == "ASC":
-            order_clause = "create_date ASC, write_date ASC, id ASC"
-
-        return self.env["mail.message"].search(domain, order=order_clause, limit=limit)
-
-    def _get_last_message_from_history(self):
-        """Get the last LLM message from the message history."""
-        self.ensure_one()
-        result = self.get_message_history_recordset(order="DESC", limit=1)
-        if not result:
-            raise UserError("No LLM message found to process.")
-        return result[0]
-
-    def _should_continue(self, last_message):
-        """Whether to keep looping based on the last message role and content."""
-        if not last_message:
-            return False
-
-        # Use the stored llm_role field for efficient checking
-        if last_message.llm_role in ('user', 'tool'):
-            return True
-
-        if last_message.llm_role == 'assistant' and last_message.tool_calls:
-            return True
-
-        return False
-
-    def get_prepend_messages(self):
-        """Hook: return a list of formatted messages to prepend to the conversation.
-        Override in other modules if needed.
-
-        Returns:
-            list: List of message dictionaries in the format:
-                [{"role": "system", "content": "..."},
-                 {"role": "user", "content": "..."},
-                 ...]
-        """
-        return []
-
-    def _execute_tool(self, tool_name, arguments_str):
-        """Execute a tool and return the result."""
-        self.ensure_one()
-        tool = self.tool_ids.filtered(lambda t: t.name == tool_name)[:1]
-        if not tool:
-            raise UserError(f"Tool '{tool_name}' not found in this thread")
-        arguments = json.loads(arguments_str)
-        return tool.execute(arguments)
+        return context
 
     # ============================================================================
     # LOCKING MECHANISM
@@ -524,23 +373,3 @@ class LLMThread(models.Model):
             self.env.user.partner_id, "llm.thread/delete", {"ids": unlink_ids}
         )
 
-    def get_context(self, base_context=None):
-        context = {
-            **(base_context or {}),
-            'thread_id': self.id,
-        }
-
-        try:
-            related_record = self.env[self.model].browse(self.res_id)
-            if related_record:
-                context['related_record'] = RelatedRecordProxy(related_record)
-                context['related_model'] = self.model
-                context['related_res_id'] = self.res_id
-            else:
-                context['related_record'] = None
-                context['related_model'] = None
-                context['related_res_id'] = None
-        except Exception as e:
-            _logger.warning("Error accessing related record %s,%s: %s", self.model, self.res_id, e)
-
-        return context
